requirements.txt
django==3.0.4
django-environ==0.4.5
django-jsonfield==1.4.0
django_multiselectfield==0.1.12
djangorestframework==3.11.0
djangorestframework_simplejwt==4.4.0

requests==2.23.0